Move DSMCC into integration interface and don't use static storage for session data
